# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bondery API
  description: >-
    REST API for the Bondery application — a contact and relationship management platform.


    ## Health checks


    - `GET /status` — **Liveness**. Returns 200 when the API process is running. Does not probe external dependencies.

    - `GET /health` — **Readiness**. Probes Supabase (auth, database, storage), Redis, and reports configuration status
    for other integrations. Cached for one minute; rate limited to one request per minute per client. Returns 503 when
    critical dependencies are unavailable.


    ## Authentication


    Most endpoints require authentication via a Supabase session cookie or a Bearer token (session JWT or long-lived API
    key).


    **Session auth:** Sign in via the webapp; the browser sends session cookies or a Supabase access token.


    **API keys:** Create keys in Settings → API keys. Send `Authorization: Bearer bondery_key_…` on allowed integration
    routes (contacts, groups, tags, interactions, imports, share, geocode). Keys support `read` (GET/HEAD) or `full`
    access. See the authentication guide.


    Endpoints under `/api/me/api-keys`, `/api/sync`, `/api/chat`, `/api/admin`, `/api/subscriptions`, and
    `/api/extension` do not accept API keys.
  version: 1.0.0
  contact:
    name: Bondery Support
    url: https://usebondery.com
  license:
    name: Proprietary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Supabase session access token from the webapp or mobile app. Send as `Authorization: Bearer <access_token>`."
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "Long-lived API key created in Settings → API keys. Format: `bondery_key_<keyId>_<secret>`. Only works on
        integration routes (contacts, groups, tags, interactions, imports, share, geocode). Supports `read` (GET/HEAD)
        or `full` access."
...
paths:
...
  /api/contacts/{id}/groups:
    get:
      tags:
        - Contacts
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
  /api/groups:
    get:
      tags:
        - Groups
      parameters:
        - schema:
            type: string
          in: query
          name: previewLimit
        - schema:
            type: string
            enum:
              - low
              - medium
              - high
          in: query
          name: avatarQuality
        - schema:
            type: string
            enum:
              - small
              - medium
              - large
          in: query
          name: avatarSize
      responses:
        "200":
          description: Default Response
    post:
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  minLength: 1
                  maxLength: 100
                emoji:
                  type: string
                  minLength: 1
                color:
                  type: string
                  pattern: ^#?[0-9a-fA-F]{6}$
              required:
                - label
                - emoji
                - color
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  minItems: 1
                  type: array
                  items:
                    type: string
              required:
                - ids
      responses:
        "200":
          description: Default Response
  /api/groups/{id}:
    get:
      tags:
        - Groups
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    patch:
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  minLength: 1
                  maxLength: 100
                emoji:
                  type: string
                  minLength: 1
                color:
                  type: string
                  pattern: ^#?[0-9a-fA-F]{6}$
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - Groups
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: http://localhost:3001
    description: Local development server
...